requirements.txt
flask==2.3.3
tensorflow==2.15.0
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
flask-cors==4.0.0
werkzeug==2.3.7